logic and ui are separated
